adjusted for decimal numbers

+++ express_calc/calc.js
@@ -4,16 +4,16 @@ app.set('view engine', 'ejs');
 
 app.get("/:op/:num1/:num2", function(req,res) {
   if(req.params.op === "add") {
-    var sum = parseInt(req.params.num1) + parseInt(req.params.num2);
+    var sum = parseFloat(req.params.num1) + parseFloat(req.params.num2);
     res.render('index',{answer: sum.toString()});
   } else if (req.params.op === "sub") {
-    var sub = parseInt(req.params.num1) - parseInt(req.params.num2);
+    var sub = parseFloat(req.params.num1) - parseFloat(req.params.num2);
     res.render('index',{answer: sub.toString()});
   } else if (req.params.op === "mult") {
-    var mult = parseInt(req.params.num1) * parseInt(req.params.num2);
+    var mult = parseFloat(req.params.num1) * parseFloat(req.params.num2);
     res.render('index',{answer: mult.toString()});
   } else if (req.params.op === "div") {
-    var div = parseInt(req.params.num1) / parseInt(req.params.num2);
+    var div = parseFloat(req.params.num1) / parseFloat(req.params.num2);
     res.render('index',{answer: div.toString()});
   } else {
     res.render('index', {answer: "That ain't right"});
